refactor: move ResourceLoader back to interfaces/ and fix export ordering
- Move ResourceLoader back to src/interfaces/ to be consistent with
  all other interface definitions in the codebase
- Accept the type-only circular dependency between interfaces/ and
  session/ as inherent to the design (TypeScript handles this fine)
- Fix alphabetical ordering in src/index.ts value and type export blocks
- Add real assertion to optional-fields test (verify thinkingLevel
  is forwarded via session.agent.state.thinkingLevel)
