+++ packages/otter-agent/src/interfaces/index.ts
@@ -13,6 +13,7 @@ export type {
 	ReadonlySessionManager,
 	EntryId,
 } from "./session-manager.js";
+export type { ResourceLoader } from "./resource-loader.js";
 export type { SkillDefinition } from "./skill-definition.js";
 export type { SkillSupportedAgentEnvironment } from "./skill-supported-agent-environment.js";
 export { isSkillSupportedAgentEnvironment } from "./skill-supported-agent-environment.js";
